Refactored Code into seperate library
