+++ gradle.properties
@@ -9,6 +9,8 @@ loader_version=0.14.19-babric.1-bta
 # HalpLibe
 halplibe_version=2.1.5
 
+prismatic_version=2.0.1
+
 # Mod
 mod_version=0.5.0
 mod_group=useless
