+++ src/main/java/useless/legacyui/Gui/Container/ContainerWorkbenchLegacy.java
@@ -14,8 +14,8 @@ import net.minecraft.core.player.inventory.slot.Slot;
 import net.minecraft.core.player.inventory.slot.SlotCrafting;
 import net.minecraft.core.world.World;
 import useless.legacyui.ConfigTranslations;
-import useless.legacyui.Gui.Slot.SlotCraftingDisplay;
-import useless.legacyui.Gui.Slot.SlotResizable;
+import useless.prismaticlibe.gui.slot.SlotCraftingDisplay;
+import useless.prismaticlibe.gui.slot.SlotResizable;
 import useless.legacyui.LegacyUI;
 import useless.legacyui.Sorting.RecipeGroup;
 import useless.legacyui.Sorting.SortingCategory;
