--- src/main/java/useless/legacyui/Gui/IButtonSounds.java
@@ -1,8 +1,0 @@
-package useless.legacyui.Gui;
-
-public interface IButtonSounds {
-
-    public boolean isMuted();
-
-    public String getSound();
-}
