--- src/main/java/useless/legacyui/Gui/Slot/IResizable.java
@@ -1,5 +1,0 @@
-package useless.legacyui.Gui.Slot;
-
-public interface IResizable {
-    public int getWidth();
-}
