--- src/main/java/useless/legacyui/Gui/Slot/SlotResizable.java
@@ -1,17 +1,0 @@
-package useless.legacyui.Gui.Slot;
-
-import net.minecraft.core.player.inventory.IInventory;
-import net.minecraft.core.player.inventory.slot.Slot;
-
-public class SlotResizable extends Slot implements IResizable{
-    public int width;
-    public SlotResizable(IInventory inventory, int id, int x, int y, int width) {
-        super(inventory, id, x, y);
-        this.width = width;
-    }
-
-    @Override
-    public int getWidth() {
-        return width;
-    }
-}
