+++ src/main/resources/legacyui.mixins.json
@@ -5,9 +5,7 @@
   "compatibilityLevel": "JAVA_8",
   "mixins": [
     "ControllerInputMixin",
-    "GuiItemDragHandlerMixin",
     "GuiOptionsBaseMixin",
-    "GuiRenderItemMixin",
     "GuiScreenMixin",
     "MinecraftMixin",
     "PlayerSPMixin"
